requirements.txt
gymnasium==0.29.1
ale-py==0.9.0
h5py==3.11.0
huggingface-hub==0.17.2
hydra-core==1.3
numpy==1.26.0
opencv-python==4.10.0.84
pillow==10.3.0
pygame==2.5.2
torch==2.1.0
torchvision==0.16.0
torcheval==0.0.7
tqdm==4.66.4
wandb==0.17.0
scikit-learn==1.5.1
scipy==1.13.1
